Add: サーバーと更新後にグラフのデータを作るように設定した
